FM2-640: added unit tests for fhir module
