+++ fhir/src/main/java/org/openmrs/module/patienttflags/dao/impl/FhirFlagDaoImpl.java
@@ -1,4 +1,13 @@
-package org.openmrs.patienttflags.dao.impl;
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.patienttflags.dao.impl;
 
 import ca.uhn.fhir.rest.param.DateRangeParam;
 import ca.uhn.fhir.rest.param.ReferenceAndListParam;
@@ -10,7 +19,7 @@ import org.openmrs.module.fhir2.FhirConstants;
 import org.openmrs.module.fhir2.api.dao.impl.BaseFhirDao;
 import org.openmrs.module.fhir2.api.search.param.SearchParameterMap;
 import org.openmrs.module.patientflags.PatientFlag;
-import org.openmrs.patienttflags.dao.FhirFlagDao;
+import org.openmrs.module.patienttflags.dao.FhirFlagDao;
 import org.springframework.stereotype.Component;
