+++ fhir/src/main/java/org/openmrs/module/patienttflags/providers/FlagFhirResourceProvider.java
@@ -1,4 +1,13 @@
-package org.openmrs.patienttflags.providers;
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.patienttflags.providers;
 
 import ca.uhn.fhir.rest.annotation.IdParam;
 import ca.uhn.fhir.rest.annotation.OptionalParam;
@@ -17,8 +26,8 @@ import org.hl7.fhir.r4.model.IdType;
 import org.hl7.fhir.r4.model.Flag;
 import org.hl7.fhir.r4.model.Patient;
 import org.openmrs.module.fhir2.api.annotations.R4Provider;
-import org.openmrs.patienttflags.FhirFlagService;
-import org.openmrs.patienttflags.search.param.FlagSearchParams;
+import org.openmrs.module.patienttflags.FhirFlagService;
+import org.openmrs.module.patienttflags.search.param.FlagSearchParams;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
