+++ fhir/src/main/java/org/openmrs/module/patienttflags/translators/PatientFlagTranslator.java
@@ -1,0 +1,27 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.patienttflags.translators;
+
+import org.hl7.fhir.r4.model.Flag;
+import org.openmrs.module.fhir2.api.translators.OpenmrsFhirTranslator;
+import org.openmrs.module.fhir2.api.translators.ToFhirTranslator;
+import org.openmrs.module.patientflags.PatientFlag;
+
+import javax.annotation.Nonnull;
+
+public interface PatientFlagTranslator extends ToFhirTranslator<PatientFlag, Flag> {
+    /**
+     * @param patientFlag 
+     * @return
+     */
+    @Override
+    Flag toFhirResource(@Nonnull PatientFlag patientFlag);
+
+}
