+++ fhir/src/main/java/org/openmrs/module/patienttflags/translators/impl/PatientFlagTranslatorImpl.java
@@ -1,4 +1,13 @@
-package org.openmrs.patienttflags.translators.impl;
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.patienttflags.translators.impl;
 
 import static org.apache.commons.lang3.Validate.notNull;
 
@@ -10,9 +19,9 @@ import org.hl7.fhir.r4.model.Flag;
 import org.hl7.fhir.r4.model.Period;
 import org.openmrs.module.fhir2.api.translators.PatientReferenceTranslator;
 import org.openmrs.module.patientflags.PatientFlag;
-import org.openmrs.patienttflags.translators.FlagTranslator;
-import org.openmrs.patienttflags.translators.PatientFlagTranslator;
-import org.openmrs.patienttflags.translators.TagTranslator;
+import org.openmrs.module.patienttflags.translators.FlagTranslator;
+import org.openmrs.module.patienttflags.translators.PatientFlagTranslator;
+import org.openmrs.module.patienttflags.translators.TagTranslator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -48,7 +57,6 @@ public class PatientFlagTranslatorImpl implements PatientFlagTranslator {
         });
 
         flag.setCode(flagTranslator.toFhirResource(patientFlag.getFlag()));
-
         flag.setSubject(patientReferenceTranslator.toFhirResource(patientFlag.getPatient()));
 
         Period period = new Period();
@@ -58,8 +66,4 @@ public class PatientFlagTranslatorImpl implements PatientFlagTranslator {
         return flag;
     }
 
-    @Override
-    public PatientFlag toOpenmrsType(@Nonnull Flag flag) {
-        return null;
-    }
 }
