+++ fhir/src/main/java/org/openmrs/module/patienttflags/translators/impl/PriorityTranslatorImpl.java
@@ -1,9 +1,18 @@
-package org.openmrs.patienttflags.translators.impl;
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.patienttflags.translators.impl;
 
 import org.hl7.fhir.r4.model.CodeableConcept;
 import org.hl7.fhir.r4.model.Coding;
 import org.openmrs.module.patientflags.Priority;
-import org.openmrs.patienttflags.translators.PriorityTranslator;
+import org.openmrs.module.patienttflags.translators.PriorityTranslator;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Nonnull;
